requirements.txt
torch==1.4.0
torchvision==0.5.0
numpy>=1.16
pillow
